pyproject.toml
[project]
name = "spreed-sql"
version = "0.1.0"
description = "SQL-like declarative schema definitions for Google Sheets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "gspread>=5.0",
    "google-auth>=2.0",
]

[project.optional-dependencies]
dev = ["pytest"]
